refactor(service): JanggiService 테스트 코드 메서드명 리팩토링
