+++ src/test/java/service/JanggiServiceTest.java
@@ -1,15 +1,14 @@
-package domain.database;
+package service;
 
 import database.context.BoardIdContext;
 import database.context.ConnectionContext;
 import database.dao.*;
 import database.dto.GameResult;
 import database.mapper.JanggiBoardMapper;
-import service.JanggiService;
+import domain.board.exception.BoardException;
 import database.SchemaInitializer;
 import domain.board.JanggiBoard;
 import database.dto.Moved;
-import domain.intersection.Intersection;
 import domain.piece.PieceType;
 import domain.piece.Team;
 import domain.point.Point;
@@ -20,7 +19,6 @@ import org.junit.jupiter.api.*;
 
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.List;
 
 import static fixture.IntersectionFixture.generate;
 
@@ -62,25 +60,9 @@ class JanggiServiceTest {
     }
 
     @Test
-    void test() throws SQLException{
-        JanggiBoard janggiBoard = JanggiBoardFixture.generate(
-                generate(0, 0, Team.CHO, PieceType.CHARIOT),
-                generate(0, 8, Team.CHO, PieceType.CHARIOT)
-        );
-
-        List<Intersection> expected = janggiBoard.getListIntersection();
-
-        // when
-        Long savedId = janggiService.createBoard(janggiBoard);
-
-        // then
-        Assertions.assertThat(intersectionDao.readByBoardId(savedId))
-                .isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("기존의 장기판 조회를 테스트한다.")
-    void test1() {
+    @DisplayName("기존의 장기게임을 불러올 때, DB로부터 저장된 기물 배치정보와 턴 정보가 일치하는 JanggiBoard를 가져온다.")
+    void shouldReturnJanggiBoardWhenGetExistBoardWithValidId() {
+        // given
         JanggiBoard expected = JanggiBoardFixture.generate(
                 generate(0, 0, Team.CHO, PieceType.CHARIOT),
                 generate(0, 8, Team.CHO, PieceType.CHARIOT)
@@ -98,8 +80,9 @@ class JanggiServiceTest {
     }
 
     @Test
-    @DisplayName("Turn을 넘겼을 때 DB에 제대로 저장되는 지 확인한다.")
-    void test2() {
+    @DisplayName("기물 이동 후, Turn을 넘겼을 때 DB에 기물 위치 변경 다음 턴의 정보가 DB에 반영된다.")
+    void shouldUpdatePiecePointAndTurnWhenUpdateTurn() {
+        // given
         Point start = new Point(0, 0);
         Point end = new Point(0, 1);
         JanggiBoard expected = JanggiBoardFixture.generate(
@@ -121,8 +104,9 @@ class JanggiServiceTest {
     }
 
     @Test
-    @DisplayName("장기의 게임 결과가 DB에 제대로 저장되는 지 확인한다.")
-    void test3() {
+    @DisplayName("장기가 종료되어, 결과를 저장할때, DB에 종료 및 승자가 반영된다.")
+    void shouldSaveGameResultWhenUpdateBoardResult() {
+        // given
         Point start = new Point(0, 0);
         Point end = new Point(0, 1);
         JanggiBoard expected = JanggiBoardFixture.generate(
@@ -140,6 +124,17 @@ class JanggiServiceTest {
         Assertions.assertThat(janggiService.getExistBoard(savedId))
                 .usingRecursiveComparison()
                 .isEqualTo(expected);
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 보드 ID로 조회 시 BoardNotFoundException이 발생한다.")
+    void shouldThrowExceptionWhenGetBoardWithInvalidId() {
+        // given
+        Long invalidId = -1L;
+
+        // when & then
+        Assertions.assertThatThrownBy(() -> janggiService.getExistBoard(invalidId))
+                .isInstanceOf(BoardException.class);
     }
 
 }
